TRADING WITH PYTHON
>> "Momentum investing" means investing in the stocks that have increased in price the most.

For this project, we build an investing strategy that selects the 50 stocks with the highest price momentum. From there, we will calculate recommended trades for an equal-weight portfolio of these 50 stocks.

>> "Value investing" means investing in the stocks that are cheapest relative to common measures of business value (like earnings or assets).

For this project, we build an investing strategy that selects the 50 stocks with the best value metrics. From there, we will calculate recommended trades for an equal-weight portfolio of these 50 stocks.